# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Burgos National High School Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /personnel:
    get:
      operationId: listPersonnel
      tags: [personnel]
      summary: List personnel
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [teaching, non_teaching, admin, null]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Personnel list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Personnel"
    post:
      operationId: createPersonnel
      tags: [personnel]
      summary: Create personnel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonnelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
  /personnel/{id}:
    get:
      operationId: getPersonnel
      tags: [personnel]
      summary: Get personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Personnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
    patch:
      operationId: updatePersonnel
      tags: [personnel]
      summary: Update personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonnelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
    delete:
      operationId: deletePersonnel
      tags: [personnel]
      summary: Delete personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── NEWS ──────────────────────────────────────────────────────────────────
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: pinned
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsListResponse"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
        "404":
          description: Not found
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ANNOUNCEMENTS ─────────────────────────────────────────────────────────
...
  /sslg:
    get:
      operationId: getSSLG
      tags: [sslg]
      summary: Get SSLG info
      responses:
        "200":
          description: SSLG info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSLG"
    put:
      operationId: updateSSLG
      tags: [sslg]
      summary: Update SSLG info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SSLGInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSLG"
  /sslg/officers:
    get:
      operationId: listSSLGOfficers
      tags: [sslg]
      summary: List SSLG officers
      responses:
        "200":
          description: Officers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SSLGOfficer"
    post:
      operationId: createSSLGOfficer
      tags: [sslg]
      summary: Create SSLG officer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SSLGOfficerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSLGOfficer"
  /sslg/officers/{id}:
    delete:
      operationId: deleteSSLGOfficer
      tags: [sslg]
      summary: Delete SSLG officer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── PTA ───────────────────────────────────────────────────────────────────
  /pta:
    get:
      operationId: getPTA
      tags: [pta]
      summary: Get PTA info
      responses:
        "200":
          description: PTA info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTA"
    put:
      operationId: updatePTA
      tags: [pta]
      summary: Update PTA info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PTAInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTA"
  /pta/officers:
    get:
      operationId: listPTAOfficers
      tags: [pta]
      summary: List PTA officers
      responses:
        "200":
          description: Officers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PTAOfficer"
    post:
      operationId: createPTAOfficer
      tags: [pta]
      summary: Create PTA officer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PTAOfficerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTAOfficer"
  /pta/officers/{id}:
    delete:
      operationId: deletePTAOfficer
      tags: [pta]
      summary: Delete PTA officer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ABOUT ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Personnel:
      type: object
      required: [id, name, position, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        position:
          type: string
        type:
          type: string
          enum: [teaching, non_teaching, admin]
        department:
          type: ["string", "null"]
        advisoryClass:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        office:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    PersonnelInput:
      type: object
      required: [name, position, type]
      properties:
        name:
          type: string
        position:
          type: string
        type:
          type: string
          enum: [teaching, non_teaching, admin]
        department:
          type: string
        advisoryClass:
          type: string
        email:
          type: string
        office:
          type: string
        photoUrl:
          type: string
    PersonnelUpdate:
      type: object
      properties:
        name:
          type: string
        position:
          type: string
        type:
          type: string
          enum: [teaching, non_teaching, admin]
        department:
          type: string
        advisoryClass:
          type: string
        email:
          type: string
        office:
          type: string
        photoUrl:
          type: string

    # ── NEWS ──────────────────────────────────────────────────────────────
    News:
      type: object
      required: [id, title, content, category, isPinned, isArchived, publishedAt, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        category:
          type: string
        isPinned:
          type: boolean
        isArchived:
          type: boolean
        coverImage:
          type: ["string", "null"]
        authorName:
          type: ["string", "null"]
        publishedAt:
          type: string
        createdAt:
          type: string
    NewsInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        category:
          type: string
        isPinned:
          type: boolean
          default: false
        isArchived:
          type: boolean
          default: false
        coverImage:
          type: string
    NewsUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        category:
          type: string
        isPinned:
          type: boolean
        isArchived:
          type: boolean
        coverImage:
          type: string
    NewsListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/News"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── ANNOUNCEMENTS ─────────────────────────────────────────────────────
...
    SSLG:
      type: object
      required: [id, adviserName, schoolYear]
      properties:
        id:
          type: integer
        adviserName:
          type: string
        schoolYear:
          type: string
        description:
          type: ["string", "null"]
        projects:
          type: ["string", "null"]
        activities:
          type: ["string", "null"]
    SSLGInput:
      type: object
      required: [adviserName, schoolYear]
      properties:
        adviserName:
          type: string
        schoolYear:
          type: string
        description:
          type: string
        projects:
          type: string
        activities:
          type: string
    SSLGOfficer:
      type: object
      required: [id, name, position, schoolYear]
      properties:
        id:
          type: integer
        name:
          type: string
        position:
          type: string
        committee:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        schoolYear:
          type: string
    SSLGOfficerInput:
      type: object
      required: [name, position, schoolYear]
      properties:
        name:
          type: string
        position:
          type: string
        committee:
          type: string
        photoUrl:
          type: string
        schoolYear:
          type: string

    # ── PTA ───────────────────────────────────────────────────────────────
    PTA:
      type: object
      required: [id, schoolYear]
      properties:
        id:
          type: integer
        schoolYear:
          type: string
        constitution:
          type: ["string", "null"]
        projects:
          type: ["string", "null"]
        activities:
          type: ["string", "null"]
        meetingSchedule:
          type: ["string", "null"]
    PTAInput:
      type: object
      required: [schoolYear]
      properties:
        schoolYear:
          type: string
        constitution:
          type: string
        projects:
          type: string
        activities:
          type: string
        meetingSchedule:
          type: string
    PTAOfficer:
      type: object
      required: [id, name, position, schoolYear]
      properties:
        id:
          type: integer
        name:
          type: string
        position:
          type: string
        schoolYear:
          type: string
    PTAOfficerInput:
      type: object
      required: [name, position, schoolYear]
      properties:
        name:
          type: string
        position:
          type: string
        schoolYear:
          type: string

    # ── ABOUT ─────────────────────────────────────────────────────────────
...